slackbot/requirements.txt
slackclient
SQLAlchemy
requests
psycopg2-binary
